Code refactor

+++ src/fuzzy/domain/DomainRange.cpp
@@ -6,26 +6,26 @@
 
 DomainRange::DomainRange(double start, double step, double end) : start(start), step(step), end(end) {}
 
-DomainElement DomainRange::getElementAt(int index) {
+DomainElement DomainRange::getElementAt(uint index) {
     if (index > getCardinality()) {
         throw std::invalid_argument("Index greater than size of domain");
     }
     return DomainElement({start + index * step});
 }
 
-int DomainRange::getCardinality() {
-    return (end - start) / step;
+uint DomainRange::getCardinality() {
+    return static_cast<int>((end - start) / step);
 }
 
-shared_ptr<Domain> DomainRange::getComponent(int index) {
+shared_ptr<Domain> DomainRange::getComponent(uint index) {
     return shared_from_this();
 }
 
-int DomainRange::indexOfElement(DomainElement element) {
+uint DomainRange::indexOfElement(DomainElement element) {
 
     return 0;
 }
 
-int DomainRange::getNumberOfComponents() {
+uint DomainRange::getNumberOfComponents() {
     return 1;
 }
